Float casting

+++ VFrame/V9Slice.cpp
@@ -42,90 +42,90 @@ void V9Slice::updateFrame()
 	vertexArray[(0 * 4) + 1].position = sf::Vector2f(cornerSize.x, 0.0f);
 	vertexArray[(0 * 4) + 2].position = sf::Vector2f(cornerSize.x, cornerSize.y);
 	vertexArray[(0 * 4) + 3].position = sf::Vector2f(0.0f, cornerSize.y);
-	vertexArray[(0 * 4) + 0].texCoords = sf::Vector2f(textureRect.left, textureRect.top);
-	vertexArray[(0 * 4) + 1].texCoords = sf::Vector2f(textureRect.left + cornerSize.x, textureRect.top);
-	vertexArray[(0 * 4) + 2].texCoords = sf::Vector2f(textureRect.left + cornerSize.x, textureRect.top + cornerSize.y);
-	vertexArray[(0 * 4) + 3].texCoords = sf::Vector2f(textureRect.left, textureRect.top + cornerSize.y);
+	vertexArray[(0 * 4) + 0].texCoords = sf::Vector2f((float)textureRect.left, (float)textureRect.top);
+	vertexArray[(0 * 4) + 1].texCoords = sf::Vector2f((float)textureRect.left + cornerSize.x, (float)textureRect.top);
+	vertexArray[(0 * 4) + 2].texCoords = sf::Vector2f((float)textureRect.left + cornerSize.x, (float)textureRect.top + cornerSize.y);
+	vertexArray[(0 * 4) + 3].texCoords = sf::Vector2f((float)textureRect.left, (float)textureRect.top + cornerSize.y);
 
 	//Top Bar
 	vertexArray[(1 * 4) + 0].position = sf::Vector2f(cornerSize.x, 0.0f);
 	vertexArray[(1 * 4) + 1].position = sf::Vector2f(cornerSize.x + barSize.x, 0.0f);
 	vertexArray[(1 * 4) + 2].position = sf::Vector2f(cornerSize.x + barSize.x, cornerSize.y);
 	vertexArray[(1 * 4) + 3].position = sf::Vector2f(cornerSize.x, cornerSize.y);
-	vertexArray[(1 * 4) + 0].texCoords = sf::Vector2f(textureRect.left + cornerSize.x, textureRect.top);
-	vertexArray[(1 * 4) + 1].texCoords = sf::Vector2f(textureRect.left + textureRect.width - cornerSize.x, textureRect.top);
-	vertexArray[(1 * 4) + 2].texCoords = sf::Vector2f(textureRect.left + textureRect.width - cornerSize.x, textureRect.top + cornerSize.y);
-	vertexArray[(1 * 4) + 3].texCoords = sf::Vector2f(textureRect.left + cornerSize.x, textureRect.top + cornerSize.y);
+	vertexArray[(1 * 4) + 0].texCoords = sf::Vector2f((float)textureRect.left + cornerSize.x, (float)textureRect.top);
+	vertexArray[(1 * 4) + 1].texCoords = sf::Vector2f((float)textureRect.left + textureRect.width - cornerSize.x, (float)textureRect.top);
+	vertexArray[(1 * 4) + 2].texCoords = sf::Vector2f((float)textureRect.left + textureRect.width - cornerSize.x, (float)textureRect.top + cornerSize.y);
+	vertexArray[(1 * 4) + 3].texCoords = sf::Vector2f((float)textureRect.left + cornerSize.x, (float)textureRect.top + cornerSize.y);
 
 	//Top Right Corner
 	vertexArray[(2 * 4) + 0].position = sf::Vector2f(Size.x - cornerSize.x, 0.0f);
 	vertexArray[(2 * 4) + 1].position = sf::Vector2f(Size.x, 0.0f);
 	vertexArray[(2 * 4) + 2].position = sf::Vector2f(Size.x, cornerSize.y);
 	vertexArray[(2 * 4) + 3].position = sf::Vector2f(Size.x - cornerSize.x, cornerSize.y);
-	vertexArray[(2 * 4) + 0].texCoords = sf::Vector2f(textureRect.left + textureRect.width - cornerSize.x, textureRect.top);
-	vertexArray[(2 * 4) + 1].texCoords = sf::Vector2f(textureRect.left + textureRect.width, textureRect.top);
-	vertexArray[(2 * 4) + 2].texCoords = sf::Vector2f(textureRect.left + textureRect.width, textureRect.top + cornerSize.y);
-	vertexArray[(2 * 4) + 3].texCoords = sf::Vector2f(textureRect.left + textureRect.width - cornerSize.x, textureRect.top + cornerSize.y);
+	vertexArray[(2 * 4) + 0].texCoords = sf::Vector2f((float)textureRect.left + textureRect.width - cornerSize.x, (float)textureRect.top);
+	vertexArray[(2 * 4) + 1].texCoords = sf::Vector2f((float)textureRect.left + textureRect.width, (float)textureRect.top);
+	vertexArray[(2 * 4) + 2].texCoords = sf::Vector2f((float)textureRect.left + textureRect.width, (float)textureRect.top + cornerSize.y);
+	vertexArray[(2 * 4) + 3].texCoords = sf::Vector2f((float)textureRect.left + textureRect.width - cornerSize.x, (float)textureRect.top + cornerSize.y);
 
 	//Left Bar
 	vertexArray[(3 * 4) + 0].position = sf::Vector2f(0.0f, cornerSize.y);
 	vertexArray[(3 * 4) + 1].position = sf::Vector2f(cornerSize.x, cornerSize.y);
 	vertexArray[(3 * 4) + 2].position = sf::Vector2f(cornerSize.x, cornerSize.y + barSize.y);
 	vertexArray[(3 * 4) + 3].position = sf::Vector2f(0.0f, cornerSize.y + barSize.y);
-	vertexArray[(3 * 4) + 0].texCoords = sf::Vector2f(textureRect.left, textureRect.top + cornerSize.y);
-	vertexArray[(3 * 4) + 1].texCoords = sf::Vector2f(textureRect.left + cornerSize.x, textureRect.top + cornerSize.y);
-	vertexArray[(3 * 4) + 2].texCoords = sf::Vector2f(textureRect.left + cornerSize.x, textureRect.top + textureRect.height - cornerSize.y);
-	vertexArray[(3 * 4) + 3].texCoords = sf::Vector2f(textureRect.left, textureRect.top + textureRect.height - cornerSize.y);
+	vertexArray[(3 * 4) + 0].texCoords = sf::Vector2f((float)textureRect.left, (float)textureRect.top + cornerSize.y);
+	vertexArray[(3 * 4) + 1].texCoords = sf::Vector2f((float)textureRect.left + cornerSize.x, (float)textureRect.top + cornerSize.y);
+	vertexArray[(3 * 4) + 2].texCoords = sf::Vector2f((float)textureRect.left + cornerSize.x, (float)textureRect.top + textureRect.height - cornerSize.y);
+	vertexArray[(3 * 4) + 3].texCoords = sf::Vector2f((float)textureRect.left, (float)textureRect.top + textureRect.height - cornerSize.y);
 
 	//Centre
 	vertexArray[(4 * 4) + 0].position = sf::Vector2f(cornerSize.x, cornerSize.y);
 	vertexArray[(4 * 4) + 1].position = sf::Vector2f(cornerSize.x + barSize.x, cornerSize.y);
 	vertexArray[(4 * 4) + 2].position = sf::Vector2f(cornerSize.x + barSize.x, cornerSize.y + barSize.y);
 	vertexArray[(4 * 4) + 3].position = sf::Vector2f(cornerSize.x, cornerSize.y + barSize.y);
-	vertexArray[(4 * 4) + 0].texCoords = sf::Vector2f(textureRect.left + cornerSize.x, textureRect.top + cornerSize.y);
-	vertexArray[(4 * 4) + 1].texCoords = sf::Vector2f(textureRect.left + textureRect.width - cornerSize.x, textureRect.top + cornerSize.y);
-	vertexArray[(4 * 4) + 2].texCoords = sf::Vector2f(textureRect.left + textureRect.width - cornerSize.x, textureRect.top + textureRect.height - cornerSize.y);
-	vertexArray[(4 * 4) + 3].texCoords = sf::Vector2f(textureRect.left + cornerSize.x, textureRect.top + textureRect.height - cornerSize.y);
+	vertexArray[(4 * 4) + 0].texCoords = sf::Vector2f((float)textureRect.left + cornerSize.x, (float)textureRect.top + cornerSize.y);
+	vertexArray[(4 * 4) + 1].texCoords = sf::Vector2f((float)textureRect.left + textureRect.width - cornerSize.x, (float)textureRect.top + cornerSize.y);
+	vertexArray[(4 * 4) + 2].texCoords = sf::Vector2f((float)textureRect.left + textureRect.width - cornerSize.x, (float)textureRect.top + textureRect.height - cornerSize.y);
+	vertexArray[(4 * 4) + 3].texCoords = sf::Vector2f((float)textureRect.left + cornerSize.x, (float)textureRect.top + textureRect.height - cornerSize.y);
 
 	//Right Bar
 	vertexArray[(5 * 4) + 0].position = sf::Vector2f(Size.x - cornerSize.x, cornerSize.y);
 	vertexArray[(5 * 4) + 1].position = sf::Vector2f(Size.x, cornerSize.y);
 	vertexArray[(5 * 4) + 2].position = sf::Vector2f(Size.x, cornerSize.y + barSize.y);
 	vertexArray[(5 * 4) + 3].position = sf::Vector2f(Size.x - cornerSize.x, cornerSize.y + barSize.y);
-	vertexArray[(5 * 4) + 0].texCoords = sf::Vector2f(textureRect.left + textureRect.width - cornerSize.x, textureRect.top + cornerSize.y);
-	vertexArray[(5 * 4) + 1].texCoords = sf::Vector2f(textureRect.left + textureRect.width, textureRect.top + cornerSize.y);
-	vertexArray[(5 * 4) + 2].texCoords = sf::Vector2f(textureRect.left + textureRect.width, textureRect.top + textureRect.height - cornerSize.y);
-	vertexArray[(5 * 4) + 3].texCoords = sf::Vector2f(textureRect.left + textureRect.width - cornerSize.x, textureRect.top + textureRect.height - cornerSize.y);
+	vertexArray[(5 * 4) + 0].texCoords = sf::Vector2f((float)textureRect.left + textureRect.width - cornerSize.x, (float)textureRect.top + cornerSize.y);
+	vertexArray[(5 * 4) + 1].texCoords = sf::Vector2f((float)textureRect.left + textureRect.width, (float)textureRect.top + cornerSize.y);
+	vertexArray[(5 * 4) + 2].texCoords = sf::Vector2f((float)textureRect.left + textureRect.width, (float)textureRect.top + textureRect.height - cornerSize.y);
+	vertexArray[(5 * 4) + 3].texCoords = sf::Vector2f((float)textureRect.left + textureRect.width - cornerSize.x, (float)textureRect.top + textureRect.height - cornerSize.y);
 
 	//Bottom Left Corner
 	vertexArray[(6 * 4) + 0].position = sf::Vector2f(0.0f, Size.y - cornerSize.y);
 	vertexArray[(6 * 4) + 1].position = sf::Vector2f(cornerSize.x, Size.y - cornerSize.y);
 	vertexArray[(6 * 4) + 2].position = sf::Vector2f(cornerSize.x, Size.y);
 	vertexArray[(6 * 4) + 3].position = sf::Vector2f(0.0f, Size.y);
-	vertexArray[(6 * 4) + 0].texCoords = sf::Vector2f(textureRect.left, textureRect.top + textureRect.height - cornerSize.y);
-	vertexArray[(6 * 4) + 1].texCoords = sf::Vector2f(textureRect.left + cornerSize.x, textureRect.top + textureRect.height - cornerSize.y);
-	vertexArray[(6 * 4) + 2].texCoords = sf::Vector2f(textureRect.left + cornerSize.x, textureRect.top + textureRect.height);
-	vertexArray[(6 * 4) + 3].texCoords = sf::Vector2f(textureRect.left, textureRect.top + textureRect.height);
+	vertexArray[(6 * 4) + 0].texCoords = sf::Vector2f((float)textureRect.left, (float)textureRect.top + textureRect.height - cornerSize.y);
+	vertexArray[(6 * 4) + 1].texCoords = sf::Vector2f((float)textureRect.left + cornerSize.x, (float)textureRect.top + textureRect.height - cornerSize.y);
+	vertexArray[(6 * 4) + 2].texCoords = sf::Vector2f((float)textureRect.left + cornerSize.x, (float)textureRect.top + textureRect.height);
+	vertexArray[(6 * 4) + 3].texCoords = sf::Vector2f((float)textureRect.left, (float)textureRect.top + textureRect.height);
 
 	//Bottom Bar
 	vertexArray[(7 * 4) + 0].position = sf::Vector2f(cornerSize.x, Size.y - cornerSize.y);
 	vertexArray[(7 * 4) + 1].position = sf::Vector2f(cornerSize.x + barSize.x, Size.y - cornerSize.y);
 	vertexArray[(7 * 4) + 2].position = sf::Vector2f(cornerSize.x + barSize.x, Size.y);
 	vertexArray[(7 * 4) + 3].position = sf::Vector2f(cornerSize.x, Size.y);
-	vertexArray[(7 * 4) + 0].texCoords = sf::Vector2f(textureRect.left + cornerSize.x, textureRect.top + textureRect.height - cornerSize.y);
-	vertexArray[(7 * 4) + 1].texCoords = sf::Vector2f(textureRect.left + textureRect.width - cornerSize.x, textureRect.top + textureRect.height - cornerSize.y);
-	vertexArray[(7 * 4) + 2].texCoords = sf::Vector2f(textureRect.left + textureRect.width - cornerSize.x, textureRect.top + textureRect.height);
-	vertexArray[(7 * 4) + 3].texCoords = sf::Vector2f(textureRect.left + cornerSize.x, textureRect.top + textureRect.height);
+	vertexArray[(7 * 4) + 0].texCoords = sf::Vector2f((float)textureRect.left + cornerSize.x, (float)textureRect.top + textureRect.height - cornerSize.y);
+	vertexArray[(7 * 4) + 1].texCoords = sf::Vector2f((float)textureRect.left + textureRect.width - cornerSize.x, (float)textureRect.top + textureRect.height - cornerSize.y);
+	vertexArray[(7 * 4) + 2].texCoords = sf::Vector2f((float)textureRect.left + textureRect.width - cornerSize.x, (float)textureRect.top + textureRect.height);
+	vertexArray[(7 * 4) + 3].texCoords = sf::Vector2f((float)textureRect.left + cornerSize.x, (float)textureRect.top + textureRect.height);
 
 	//Bottom Right Corner
 	vertexArray[(8 * 4) + 0].position = sf::Vector2f(Size.x - cornerSize.x, Size.y - cornerSize.y);
 	vertexArray[(8 * 4) + 1].position = sf::Vector2f(Size.x, Size.y - cornerSize.y);
 	vertexArray[(8 * 4) + 2].position = sf::Vector2f(Size.x, Size.y);
 	vertexArray[(8 * 4) + 3].position = sf::Vector2f(Size.x - cornerSize.x, Size.y);
-	vertexArray[(8 * 4) + 0].texCoords = sf::Vector2f(textureRect.left + textureRect.width - cornerSize.x, textureRect.top + textureRect.height - cornerSize.y);
-	vertexArray[(8 * 4) + 1].texCoords = sf::Vector2f(textureRect.left + textureRect.width, textureRect.top + textureRect.height - cornerSize.y);
-	vertexArray[(8 * 4) + 2].texCoords = sf::Vector2f(textureRect.left + textureRect.width, textureRect.top + textureRect.height);
-	vertexArray[(8 * 4) + 3].texCoords = sf::Vector2f(textureRect.left + textureRect.width - cornerSize.x, textureRect.top + textureRect.height);
+	vertexArray[(8 * 4) + 0].texCoords = sf::Vector2f((float)textureRect.left + textureRect.width - cornerSize.x, (float)textureRect.top + textureRect.height - cornerSize.y);
+	vertexArray[(8 * 4) + 1].texCoords = sf::Vector2f((float)textureRect.left + textureRect.width, (float)textureRect.top + textureRect.height - cornerSize.y);
+	vertexArray[(8 * 4) + 2].texCoords = sf::Vector2f((float)textureRect.left + textureRect.width, (float)textureRect.top + textureRect.height);
+	vertexArray[(8 * 4) + 3].texCoords = sf::Vector2f((float)textureRect.left + textureRect.width - cornerSize.x, (float)textureRect.top + textureRect.height);
 }
 
 V9Slice::V9Slice(sf::Vector2f Position, sf::Vector2f Size, const sf::String& Filename, const sf::Vector2f& Corner) : VObject(Position, Size)
